captureLogics in its own file
